+++ app/src/main/java/com/jin/honey/feature/payment/domain/usecase/GetOrderHistoryUseCase.kt
@@ -4,7 +4,7 @@ import com.jin.honey.feature.payment.domain.PaymentRepository
 import com.jin.honey.feature.payment.domain.model.Payment
 
 class GetOrderHistoryUseCase(private val repository: PaymentRepository) {
-    suspend operator fun invoke(): List<Payment> {
-        return repository.findOrderHistory()
+    suspend operator fun invoke(): Result<List<Payment>> {
+        return repository.fetchOrderHistory()
     }
 }
